requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
flask
torch
torchvision
Pillow
requests
python-dotenv
gunicorn